desktop: route chat-tab focus through the arbiter, trim MET-152 tests
The composer was self-focusing (PromptEditor autoFocus + a direct
editorRef.focus() when the session finished loading), so agent tabs were the
one tab type whose focus bypassed arbitration — it could yank a modal, menu
or sidebar text entry that outranks an ambient tab intent. It now files a
when-mounted intent like a document editor's mount does, and the controller
resolves it into the composer's own focus call.

Tests: 29 new cases down to 13. Dropped the tab-id scheme duplicate (the
entity test already covers it), folded the search/history cases into one
editor-tab-controller test, and cut assertions that only restated a stub.
